Add new ship selection dialog.  Add new DragButton.  Rework tweens to remove autostart.

+++ GalCon-Client/GalCon-Client/src/com/xxx/galcon/screen/MainMenuScreen.java
@@ -21,7 +21,6 @@ import com.xxx.galcon.GameLoop;
 import com.xxx.galcon.InGameInputProcessor;
 import com.xxx.galcon.InGameInputProcessor.TouchPoint;
 import com.xxx.galcon.ScreenFeedback;
-import com.xxx.galcon.http.ConnectionException;
 import com.xxx.galcon.http.GameAction;
 import com.xxx.galcon.http.SetPlayerResultHandler;
 
@@ -95,7 +94,7 @@ public class MainMenuScreen implements ScreenFeedback {
 		font.draw(spriteBatch, text, x, y);
 
 		if (isTouchable && touchRegions.size() != 3) {
-			touchRegions.put(text, new TouchRegion(x, y, fontBounds.width, fontBounds.height, true));
+			touchRegions.put(text, new TouchRegion(x, y, (int) fontBounds.width, (int) fontBounds.height, true));
 		}
 	}
 
@@ -123,7 +122,6 @@ public class MainMenuScreen implements ScreenFeedback {
 		spriteBatch.setProjectionMatrix(viewMatrix);
 		spriteBatch.setTransformMatrix(transformMatrix);
 		spriteBatch.begin();
-		
 
 		String galcon = "GalCon";
 		BitmapFont extraLargeFont = Fonts.getInstance().largeFont();
@@ -132,9 +130,9 @@ public class MainMenuScreen implements ScreenFeedback {
 
 		BitmapFont smallFont = Fonts.getInstance().smallFont();
 		if (hasUserInformation()) {
-			
+
 			createCoinDisplay(width, height);
-			
+
 			BitmapFont mediumFont = Fonts.getInstance().mediumFont();
 			mediumFont.setColor(0.2f, 1.0f, 0.2f, 1.0f);
 			smallFont.setColor(0.2f, 1.0f, 0.2f, 1.0f);
@@ -191,7 +189,7 @@ public class MainMenuScreen implements ScreenFeedback {
 		String coinsText = "" + GameLoop.USER.coins;
 		BitmapFont extraLargeFont = Fonts.getInstance().largeFont();
 		double percentageOfWidth = width * 0.04;
-		int x = (int)percentageOfWidth;
+		int x = (int) percentageOfWidth;
 		extraLargeFont.draw(spriteBatch, coinsText, x, (int) (height * .97f));
 	}
